Do not enforce unique version for component versions which need be overrideable
In particular the Hibernate 2LC modules are also included in WildFly, however
when the server is provisioned we should be using the version provided by WildFly
as e.g. the "modules for WildFly" are meant to be tested with exactly what is
provided in a given version of the WildFly server.